Ranking service

+++ services/ranking-service/src/main/java/com/dinetime/ranking/presentation/request/LeaderboardsRequestModel.java
@@ -18,6 +18,6 @@ public class LeaderboardsRequestModel {
     @NoArgsConstructor
     public static class LeaderboardEntry {
         private int mealId;
-        private double totalScore;
+        private int totalScore;
     }
 }
